refactor(mcp): remove auto-start, require Appium to be running externally
